+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- [AdSupport custom Jack-in command lines](https://github.com/BetterThanTomorrow/calva/issues/2152)
 - [Add docs about main-opts and make the warning link there](https://github.com/BetterThanTomorrow/calva/issues/2157)
 
 ## [2.0.351] - 2023-04-12